requirements.txt
alembic==1.13.3
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
SQLAlchemy-serializer==1.4.12
typing_extensions==4.12.2
virtualenv==20.26.3
Werkzeug==3.0.4
zipp==3.20.2